move interceptForCrossClusterAccessRequests and RCS_INTERNAL_ACTIONS_REPLACEMENTS
